test: integration test script without using pytest
