+++ pyproject.toml
@@ -39,7 +39,7 @@ target-version = ["py311", "py312"]
 
 [tool.pytest.ini_options]
 # TODO: re-add test_traces.py when we have a better way to support it in CI
-addopts = "--ignore=tests/lib --ignore=tests/test_traces.py"
+addopts = "--ignore=tests/lib --ignore=tests/test_halmos.py --ignore=tests/test_traces.py"
 # addopts = "--ignore=tests/lib"
 
 [tool.ruff.lint]
